+++ HISTORY.rst
@@ -31,6 +31,7 @@ latest (unreleased)
 * Country date formats
 * Do not install PLM and Quality modules
 * Zendesk ticket numbers validation
+* BSSFL-306: Create a Sensefly state on sale order
 
 **Bugfixes**